third_party/WebKit/LayoutTests/http/tests/devtools/console-cross-origin-iframe-logging-expected.txt

CONSOLE ERROR: line 19: Failed to execute 'postMessage' on 'DOMWindow': The target origin provided ('http://127.0.0.1:8000') does not match the recipient window's origin ('http://localhost:8000').
Tests that cross origin errors are logged with source url and line number.


console-cross-origin…ame-logging.html:19 Failed to execute 'postMessage' on 'DOMWindow': The target origin provided ('http://127.0.0.1:8000') does not match the recipient window's origin ('http://localhost:8000').
accessFrame @ console-cross-origin…ame-logging.html:19
(anonymous) @ VM:1
